radiomics_extraction/pyproject.toml
[project]
name = "radiomics-extraction"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "numpy",
    "SimpleITK",
    "pyradiomics",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
